ai_models/requirements.txt
flask
numpy
pillow
tensorflow
keras
flask-cors
